Improve type inference
Summary:
Previously, it was impossible to create Dagster types directly wrapping
built-ins and extensions, because we attempted to monkeypatch the Python type
with an annotation pointing to the Dagster type. Additionally, attempting to
annotate a solid with a Python type outside of a narrow set would result in an
unexpected error, if that type were not explicitly wrapped in as_dagster_type or
equivalent. This diff replaces the monkeypatching machinery with an explicit
runtime type registry, and also adds a facility for automatically wrapping bare
Python types in anonymous Dagster types.

Test Plan: Unit

Differential Revision: https://dagster.phacility.com/D1273