(intermediate-storage-instance-config-1) Eliminate FilesystemIntermediateStore subclass
Summary: This is the beginning of a sequence whose goal is to make the
intermediates storage configurable at the instance level while also
making that config overrideable at the run-level.

However along that way I plan on making improvements to the plugin
system and some of our architecture.

First is eliminating the FilesystemIntermediateStore entirely. This can be
dealt with via composition, which is nicer. Also this eliminates the
dependency on DagsterInstance, which is unnecessary.

Test Plan: BK

Reviewers: alangenfeld, max, #ft

Differential Revision: https://dagster.phacility.com/D1304